+++ R/include_shape.R
@@ -2,64 +2,89 @@
 #'
 #' Insert shape in label
 #'
-#' @param label label output
-#' @param value type of shape
-#' @param units units for shape
-#' @param size image size
-#' @param position position coordinate
-#' @param border_color image color
-#' @param border_width shape line width
-#' @param background background color
-#' @param margin = shape margin
-#' @param value column or path
-#' @param panel_color panel color
-#' @param panel_size panel size
+#' @param label label output (table)
+#' @param value type of shape (string: "hexagon")
+#' @param size shape size (numeric: 5.08)
+#' @param position position coordinate (numeric: NA)
+#' @param border_color image color (string: "black")
+#' @param border_width shape line width (numeric: 1)
+#' @param background background color (string: "red")
+#' @param units units for shape (string: "cm")
+#' @param panel_color panel color (string: NA)
+#' @param panel_size panel size (numeric: NA)
 #'
 #' @return data frame
 #'
 #' @export
 #'
 #' @examples
-#'
+#' 
+#' library(huito)
 #'
 #' label <- label_layout(data = NA
 #'                , size = c(10, 2.5)
+#'                , background = "yellow"
 #'                ) %>%
 #'          include_shape(
 #'                value = "hexagon"
 #'                , position = c(1.2, 1.25)
-#'                , size = 3
+#'                , background = "red"
+#'                , border_width = 1
+#'                , size = 2.4
+#'                #, panel_size = 2.4*1.157175
 #'                )
 #'                
 #' label %>% label_print("sample")
 #' 
 #' ts <- label$opts
 #'                
-#'
 
 include_shape <- function(label
                   , value = "hexagon"
-                  , size = 4.325
+                  , size = 5.08
                   , position = NA
                   , border_color = "black"
-                  , border_width = 2
+                  , border_width = 1
                   , background = NA
-                  , margin = 0
                   , units = "cm"
                   , panel_color = NA
-                  , panel_size = 5.08
+                  , panel_size = NA
                   ) {
   
+  if (FALSE) {
+    
+    value = "hexagon"
+    size = 2.5
+    position = NA
+    border_color = "black"
+    border_width = 1
+    background = NA
+    units = "cm"
+    panel_color = NA
+    panel_size = 5.08
+    
+  }
+  
 # param -------------------------------------------------------------------
-
+  
+  panel_size <- if(any(is.null(panel_size)) || any(is.na(panel_size)) || any(panel_size == "")) {
+    size*1.157175
+  } else if(is.character(panel_size)) {
+    panel_size %>% as.numeric()
+  } else { panel_size }
+  
   size <- if(any(is.null(size)) || any(is.na(size)) || any(size == "")) {
-    4.325
+    stop("include hexagon size")
   } else if(is.character(size)) {
     size %>%
       gsub("[[:space:]]", "", .) %>%
       strsplit(., "[*]") %>%
       unlist() %>% as.numeric()
-  } else {size}
+  } else { size }
+  
+  if(size > panel_size) "Panel size should be bigger than the hexagon size"
+  
+# -------------------------------------------------------------------------
 
   position <- if(any(is.null(position)) || any(is.na(position)) || any(position == "")) {
     c(0, 0)
@@ -86,29 +111,10 @@ include_shape <- function(label
     "transparent"
   } else {border_color}
   
-  margin <- if(any(is.null(margin)) || any(is.na(margin)) || 
-               any(margin == "") ) {
-    rep(0, times = 4)
-  } else if(is.character(margin)) {
-    margin %>%
-      gsub("[[:space:]]", "", .) %>%
-      strsplit(., "[*]") %>%
-      unlist() %>% as.numeric()
-  } else if (length(margin) == 1 && is.numeric(margin)) {
-    rep(margin, times = 4)
-  } else {margin}
-  
-  
   panel_color <- if(any(is.null(panel_color)) || any(is.na(panel_color)) || any(panel_color == "")) {
     "transparent"
   } else {panel_color}
   
-  panel_size <- if(any(is.null(panel_size)) || any(is.na(panel_size)) || any(panel_size == "")) {
-    5.08
-  } else if(is.character(panel_size)) {
-    panel_size %>% as.numeric()
-  } else {panel_size}
-
 # options -----------------------------------------------------------------
 
   opt <- list(value = value
@@ -117,7 +123,6 @@ include_shape <- function(label
               , color = background
               , border_color = border_color
               , border_width = border_width
-              , margin = margin %>% paste0(collapse = "*")
               , units = units
               , panel_color = panel_color
               , panel_size = panel_size
@@ -132,7 +137,7 @@ include_shape <- function(label
       , TRUE ~ nlayer
     ))
 
-  # result ------------------------------------------------------------------
+# result ------------------------------------------------------------------
 
   result <- list(opts = opt
                  , data = label$data)